pyproject.toml
[tool.poetry]
name = "lolrmm"
version = "0.1.0"
description = ""
authors = ["Magic Sword <info@magicsword.io>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "1.54.0"
pyyaml = "^6.0.3"
jinja2 = "^3.1.4"
jsonschema = "^4.23.0"
ruamel-yaml = "^0.19.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
